+++ CHANGELOG.md
@@ -17,7 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Changed
 
-- Redesigned drivers to inherit routine code from a parent `Driver` class.
+- Redesigned drivers and workers to use `Driver` and `tasks` to reduce redundant calculations.
 - Drivers and workers should only return guaranteed values.
 `DriverOpt` and respective workers no longer return gradients.
 - Renamed `xtb_engrad` to `xtb_python_engrad` to reflect the Python interface being used.